Filtering/requirements.txt
open3d>=0.18.0
pyrealsense2>=2.54.1
numpy>=1.23
pyyaml>=6.0